backend/requirements.txt
Django
mysqlclient
djangorestframework
django-cors-headers
channels
channels_redis

